# [Enhancement] optimize distinct agg