… test/sql/test_agg/R/test_distinct_agg
… test/sql/test_agg/T/test_distinct_agg