Fixes #25641

## Problem
According to the problems reported by clients, a similar test was carried out on the development machine, and the results are as follows:
```
mysql>  select count(distinct c2), count(distinct c5), count(*) from skew_agg;
+--------------------+--------------------+----------+
| count(DISTINCT c2) | count(DISTINCT c5) | count(*) |
+--------------------+--------------------+----------+
|                  8 |           30460361 | 39359561 |
+--------------------+--------------------+----------+

mysql>   select count(distinct c1,c5), count(c4) from skew_agg group by c2;
8 rows in set (1 min 22.50 sec)

mysql> select count(*), sum(cnt) from (select c1, c5, c2, count(c4) cnt from skew_agg group by c1, c2,c5) t group by c2;
8 rows in set (10.23 sec)

mysql> explain logical select count(distinct c1,c5), count(c4) from skew_agg group by c2;
+----------------------------------------------------------------------------------------------------------------------------+
| Explain String                                                                                                             |
+----------------------------------------------------------------------------------------------------------------------------+
| - Output => [7:count, 8:count]                                                                                             |
|     - AGGREGATE(GLOBAL) [3:c2] => [7:count, 8:count]                                                                       |
|         - AGGREGATE(DISTINCT_GLOBAL) [2:c1, 3:c2, 6:c5]                                                                    |
|             - EXCHANGE(SHUFFLE) [3]                                                                                  |
|                 - AGGREGATE(LOCAL) [2:c1, 3:c2, 6:c5]                                                                      |
|                     - SCAN [skew_agg] => [2:c1, 3:c2, 6:c5, 9:c4]                                                          |
+----------------------------------------------------------------------------------------------------------------------------+
16 rows in set (0.00 sec)
```
Due to c2 is a low cardinality col, the DISTINCT_GLOBAL aggregation process cannot fully leverage the advantages of parallel computing.
Besides, c5 is a high cardinality col, it leads to a much lower efficiency in the final aggregation result compared to splitting it into two stages.

The Skew hint can solve similar problems in certain scenarios, such as `SELECT COUNT(DISTINCT c5), c2 FROM skew_agg GROUP BY c2`. However, 
it cannot handle distinct multiple cols or select list contains other aggregate functions.

## Main work
This PR optimizes the execution plan of the original three-stage aggregation by modifying the SplitAggregateRule. 
The main modifications are as follows:

1. Set the `partitionByCols` of `DISTINCT_GLOBAL` when necessary. If the cardinality of the original `groupBy keys` is high enough for parallel processing, use the `groupBy keys`; otherwise, use group by `key + distinct columns`. When there are no statistics available, the default is to use group by key + distinct columns.
2. If use grouping keys and distinct cols to distribute data, we need to continue split the final global agg.
3. The new plan may exist `LOCAL` agg with distinct agg func in update phase but other agg func in merge phase, the corresponding properties need to be set correctly in the build fragment phase.

After this PR:
```
mysql>  select count(distinct c1,c5), count(c4) from skew_agg group by c2;
8 rows in set (11.81 sec)

mysql> explain logical select count(distinct c1,c5), count(c4) from skew_agg group by c2;
+------------------------------------------------------------------------------------------------------------------------------------+
| Explain String                                                                                                                     |
+------------------------------------------------------------------------------------------------------------------------------------+
| - Output => [7:count, 8:count]                                                                                                     |
|     - AGGREGATE(GLOBAL) [3:c2] => [7:count, 8:count]                                                                               |
|         - EXCHANGE(SHUFFLE) [3]                                                                                                    |
|             - AGGREGATE(LOCAL) [3:c2]                                                                                              |
|                 - AGGREGATE(DISTINCT_GLOBAL) [2:c1, 3:c2, 6:c5]                                                                    |
|                     - EXCHANGE(SHUFFLE) [2, 3, 6]                                                                                  |
|                         - AGGREGATE(LOCAL) [2:c1, 3:c2, 6:c5]                                                                      |
|                             - SCAN [skew_agg] => [2:c1, 3:c2, 6:c5, 9:c4]                                                          |
+------------------------------------------------------------------------------------------------------------------------------------+
```



## What type of PR is this:
- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

## Checklist:
- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5
  - [ ] 2.4