@wanpengfei-git 2023-06-27
#### [FE PR Coverage Check]

:heart_eyes: **pass** : 45 / 46 (97.83%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/sql/parser/AstBuilder.java|1|2|50.00%|[5568]|
|:large_blue_circle:|com/starrocks/sql/optimizer/validate/TypeChecker.java|8|8|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/transformation/SplitAggregateRule.java|28|28|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/plan/PlanFragmentBuilder.java|8|8|100.00%|[]|