@Seaven 2023-06-28
we have tree rule to optimize sepecial aggregate, we can move there rule to a special transform rule in CBO，and don't split aggregate node to multi-phase？

`SplitAggregate` is too complex now