@Seaven 2023-06-29
> Do you
> 
> > we have tree rule to optimize sepecial aggregate, we can move there rule to a special transform rule in CBO，and don't split aggregate node to multi-phase？
> > `SplitAggregate` is too complex now
> 
> Do you mean use special rules to process different `distinct agg` scenes?

Yeah, sr has `skew`/`RewriteCountDistinct` logical rule to resolve `distinct agg`, they are depend on special cardinality or cost, and split to multi-agg, it's similar to your optimize. I think we can support a speical rule to resolve it in MEMO, not case by case to handle