@liuyehcf 2023-07-03
> Due to c2 is a low cardinality col, the DISTINCT_GLOBAL aggregation process cannot fully leverage the advantages of parallel computing.

I think the actual reason is data_skew but not cardinality? Both high or low cardinality with skewed can lead to poor performance.